Attempt to make the OCR more precise
In my tests, the OCR has better times with PNG (probably due to
compression). I did not manage to get a lossless format working so far,
but should be a good thing to do in the future.

Also, whitelisting the symbols (degrees, minutes, seconds), gives some
hints to the parser and makes it more precise.