Introduce the concept of Regions
It turns out we have way too many raster images, specially on big
cities. A lot of these images are hidden and are never rendered. This
patch is the early bits of a trivial scene graph implementation.

The major goal will be to render cities by regions of clamped images and
do not add to the .vrt images that are completely occluded.