Add occlusion culling
A trivial hit test to discard images that are not exactly visible.
On the next step I will add image grouping so we won't do the hit test
on the empty areas of the map.

Testing with Para de Minas, half of the images were discarded and I did
not notice any difference on the final result. \o/