Group images by Regions
This is outstanding optimization. Usually a city has a lot of other
districts in remote areas. If we consider everything as a single map, we do
a lot of computation in the areas within these districts, and they are
just empty spaces.

In the future patches adding support for .mbtiles, we are going to do
exports by Regions. A single city like Para de Minas, will generate 8
.mbtiles and thus, 8 data exports. The performance gain is massive,
since they can be done in parallel and we do not generate transparent
tiles for the areas in between.