Add the option for generating tiles by regions
This wasn't really a problem for the southeast of Brazil, but now in the
north we have sparse cities like Altamira (~160k km2). This will take a
lot of time to process because of the empty areas. But we can devide
the city into "PDF agglomerations" (aka regions) and merge them after
processing if needed.