Parse cities at the Equator line
Duh, I made the silly assumption that I could return 0 degrees for
invalid coordinates when running the OCR and completely ignored the fact
that we have a lot of cities, notably Macapa, at the Equator.

Also decreased the OCR confidence to 50, since that is what I get when
parsing cities with coordinates with a lot of zeroes (my guess is the
OCR ranks 0 lower because it looks a lot like O).